+++ src/pages/jia/componentsJia/ArticleJia/ArticleJia.scss
@@ -119,10 +119,109 @@ article {
 
   .feedsMoreComments {
     margin-top: 10px;
+    font-weight: 500;
+    color: #bababa;
+  }
+}
 
-    .feedsMoreCommentsText {
-      font-weight: 500;
+.feedsCommentBox {
+  font-size: 14px;
+  border-bottom: 1px solid #e6e6e6;
+
+  .strong {
+    font-weight: 700;
+  }
+
+  .feedsOldComment {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    margin-top: 10px;
+    width: 100%;
+
+    div:first-child {
+      width: 100%;
+      display: flex;
+      justify-content: space-between;
+      align-items: center;
+    }
+
+    .feedsOldCommentMore {
+      color: gray;
+      font-size: 14px;
+      font-weight: 400;
+      margin-top: 10px;
+    }
+
+    .heart-icon-mini {
+      width: 10px;
+      height: 10px;
+    }
+  }
+
+  .feedsNewComment {
+    display: flex;
+    flex-direction: column;
+    justify-content: space-between;
+
+    .feedsCommentInputBox {
+      display: flex;
+      justify-content: space-between;
+      align-items: center;
+      height: 40px;
+      padding-bottom: 5px;
       color: #bababa;
+      font-size: 14px;
+
+      .feedsCommentInput {
+        margin: 0;
+        padding: 0;
+        width: 90%;
+        background-color: white;
+        border: none;
+        font-size: 14px;
+
+        &::placeholder {
+          color: #bababa;
+          text-align: left;
+        }
+      }
+
+      .feedsCommentUpload {
+        padding: 0;
+        width: 10%;
+        color: #5395e9;
+        background-color: transparent;
+        border: none;
+        text-align: right;
+      }
+    }
+
+    li {
+      margin-bottom: 6px;
+    }
+
+    span {
+      font-weight: 700;
+      margin-right: 5px;
+    }
+
+    input {
+      width: 180px;
+      height: 20px;
+      padding-left: 10px;
+      background-color: #fafafa;
+      border: none;
+      border-radius: 4px;
+
+      &::placeholder {
+        color: #bababa;
+        text-align: center;
+      }
+
+      &:focus {
+        outline: none;
+      }
     }
   }
 }
