+++ src/pages/jia/componentsJia/NavJia/NavJia.scss
@@ -46,6 +46,7 @@ button {
   padding: 40px 120px 40px 25px;
   border-right: 1px solid #e6e6e6;
   background-color: white;
+  z-index: 1;
 
   img {
     width: 25px;
